+++ src/Products/PloneMeeting/browser/overrides.py
@@ -58,6 +58,7 @@ from Products.CMFPlone.utils import safe_unicode
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.PloneMeeting import utils as pm_utils
 from Products.PloneMeeting.config import BARCODE_INSERTED_ATTR_ID
+from Products.PloneMeeting.config import HAS_RESTAPI
 from Products.PloneMeeting.config import ITEM_DEFAULT_TEMPLATE_ID
 from Products.PloneMeeting.config import ITEM_SCAN_ID_NAME
 from Products.PloneMeeting.config import MEETINGMANAGERS_GROUP_SUFFIX
@@ -397,6 +398,13 @@ class PloneMeetingOverviewControlPanel(OverviewControlPanel):
             versions.insert(1, "%s %s (%s)" % (plugin_package_name.split('.')[1],
                                                plugin_version,
                                                plugin_ps_version))
+        # WS, imio.pm.ws
+        ws_soap_version = api.env.get_distribution('imio.pm.ws')._version
+        versions.insert(2, 'imio.pm.ws %s' % ws_soap_version)
+        # plonemeeting.restapi could be not found on older versions
+        if HAS_RESTAPI:
+            ws_rest_version = api.env.get_distribution('plonemeeting.restapi')._version
+            versions.insert(2, 'plonemeeting.restapi %s' % ws_rest_version)
         return versions